Make Account parent relation nullable
The root account does not have a parent.  I'm not sure what would have
happened when trying to get the root account's parent, but this change
fixes blank transaction lists due to select_related() not returning any
rows for accounts that were only nested a couple of levels under the
root account.